pyproject.toml
[tool.poetry]
name = "duckdb-etl"
version = "0.1.0"
description = ""
authors = ["gabrielgarciae <gabrielgevangelista@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gdown = "^5.1.0"
duckdb = "^0.10.0"
streamlit = "^1.32.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
